fixed memory leaks
Started work on file transfer.

+++ RemoteDesktopServer/stdafx.h
@@ -41,7 +41,7 @@ int ci_find_substr(const T& str1, const T& str2, const std::locale& loc = std::l
 {
 	typename T::const_iterator it = std::search(str1.begin(), str1.end(),
 		str2.begin(), str2.end(), my_equal<typename T::value_type>(loc));
-	if (it != str1.end()) return it - str1.begin();
+	if (it != str1.end()) return (int)(it - str1.begin());
 	else return -1; // not found
 }
 #if _DEBUG
